feat: :sparkles: Seccion3 complete
