+++ src/router/index.js
@@ -18,6 +18,11 @@ const routes = [
     component: () => import(/* webpackChunkName: "about" */ '../views/BlogView.vue')
   },
   {
+    path: '/seccion3',
+    name: 'seccion3',
+    component: () => import(/* webpackChunkName: "about" */ '../views/Seccion3View.vue')
+  },
+  {
     path: '/blog/:id',
     name: 'article',
     component: () => import(/* webpackChunkName: "about" */ '../views/ArticleView.vue')
